Remove thread name from conversion pattern
The thread name in the log4j appender sample conversion pattern should be removed. The example provided does not work with kafka since the log pipeline fails to parse the pattern. Because of this datadog fails to extract the predefined facets such as timestamp and log level from the log entry.